project_18_deploy_agent/requirements.txt
google-adk==0.1.0
